src/requirements.txt
# pip install -r requirements.txt -f https://download.pytorch.org/whl/torch_stable.html

colorama==0.4.3
coloredlogs==14.0
imageio==2.8.0
numpy==1.18.5
opencv-python==4.2.0.34
pillow==7.1.2
requests==2.23.0
rsa==4.6
torch===1.5.0
torchvision===0.6.0
watchdog==0.9.0 # 0.10.2
